Adding a new field to the summaries for the Search page
I used the Standard Redactor in Craft to allow a tiny bit more of HTML
in a new kind of ‘body’ field used on some pages.  I did this so that I
could prevent some of the accordion HTML code from getting erased.  I
was able to get the JQuery/JavaScript for the Accordion widget into a
template, but the HTML still needs to be entered into a form field, for
now.