Created a Medium Blue color for links
I was getting so many complaints from people that the light blue color
was impossible to see.  The buttons were fine, but the color didn’t
translate well to links.  I darkened the links ever so slightly that I
think many people will not notice, but at least the links are easier to
see, now.