General fixes and enhancements
I fixed a pagination bug on the Media page.  I might re-work it so that
a different pagination partial template is not needed for Media.  I
changed a display area for messages on subPage.html and I was working
with live.html to accommodate our new streaming service and decided to
hide the title for that template.