Removing the item date from the display of this feed
The City’s pubDate is the creation date of the post instead of the
Event Date, which doesn’t make it useful for display on our website.  A
help ticket has been sent to The City to see if they will fix this. If
not, I will just use it without displaying the date here.  The event
dates will display when visitors click the Read More links.