Adding a feature to hide the top navigation when needed.
In an effort to reduce duplicated pages of content for the website and
St. Anne phone app, I added a JavaScript/UI feature that enables pages
to look like they are part of the phone app even though they are pages
from the website.  To use it, add ‘?app=1’ to the end of any URL from
the website that is going into the phone app.