Using different feeds in our website
News & Updates was getting difficult to maintain, because we don’t have
enough people willing to write news articles.  I’ve replaced it with
Announcements, for now, but I don’t want to disable all the previous
News items.  I removed the News Landing page from the main menu and
made a version of it in a different area that doesn’t show up in
searches on the website.