Adding a different menu behavior for Ministries and Movements
Many of us felt that the Ministries and Movements were buried too deep
in the menu, but it is necessary for our leadership that they appear to
be children of “Community of Communities” in some way.  Ideally, we
don’t want people to bypass “Community of Communities” when they are
looking and ministries and movements, but it is clear from the web
analytics that people aren’t even visiting this page, anyway.  None of
the ministry and movement pages are getting any visits, because people
can’t find them.