Removing the Slugify plugin.
It appears to have only been used on this one template.  The plugin
isn’t available in Craft 3, yet.  Most importantly, when I removed the
plugin and reference to it in this template and reloaded the page, the
rendering and class name in the code did not seem to change.