Performance improvement - moving more fields out of the Globals area
I added Mass/Confession Times fields to a Channel section, because I
want our priest secretary to be able to edit the schedule on the
website. I also didn’t make it a Single, because I want the page to
show up in the regular menu structure without needing a redirect.