Removed version from folder name
We're using Craft 4 and may update to Craft 5 in the near future.  So, there is no need to have "3" in the folder name.  Also, the staging and live sites don't have 3 in the folder name anymore.  Removing it will enable files to be copied to the correct location when deployed.